New tiles and player sprite
Put the new tiles in, did a few test maps, updated the base map template to have light and resized it and stuff, and changed the player sprite. I think that's most everything... I organized the tiles folder a bit and renamed the old tilepallets so they would appear last on the list.